improve brushed bar chart. Fix text alignment, brush updates and axis
