+++ neuralqa/server/ui/src/components/explainview/explainview.css
@@ -11,11 +11,6 @@
   overflow: scroll;
 }
 
-.barrect {
-  stroke-width: 1px;
-  stroke: rgb(212, 212, 212);
-  /* fill: #0062ff; */
-}
 .viztabcontent {
   background-color: rgb(235, 235, 235);
 }
